Implemented message-character limits in tcoin.cpp and pcoin.cpp
After having received a 10000-character transaction message from ~jan6@envs.net,
I finally came around to implementing character limits on transaction messages.
Previously, I was of the opinion that one should be able to send a movie as a
transaction message if one so wished.
I took inspiration for an appropriate character limit from Australia's NPP (New
Payments Platform), which has a 280-character limit on transaction messages. I
strongly suspect this limit came from Twitter's analysis of tweets in different
languages and Twitter research that concluded that a 280-character tweet was the
appropriate length of tweet. A transaction message is supposedly like a tweet with
some money sent along with it.