Made a minor mistake in pcoin.cpp regarding symbolic constant TCOIN_MSG_LENGTH_LIMIT
I made a switch from TCOIN_MSG_LENGTH to TCOIN_MSG_LENGTH_LIMIT for the symbolic constant
representing the maximum number of characters of a transaction message. I changed all instances
of the symbolic constant to this new name in tcoin.cpp, but forgot to do it in one instance in
pcoin.cpp. This goes on to show that I need to develop tests for tcoin and pcoin, as well as
start using getopt from util-linux for proper options handling.