Fixed a mistake in tcoin --help
Brought to my attention by ~mastergeek on tilde.team, the tcoin help incorrectly said that the number of tildecoins a user started with was 1000. It was actually the current base amount (unaltered by other sources like minercoin, tildebot or daily-adventure). Now, the help shows the right amount of starting coins.