Fixed a small syntax error in tcoin.cpp
I was planning to make a change that I did not adequately reverse, so I have reversed it fully now.
The change I was planning to make was to remove references to login@tilde.town on systems that were
not tilde.town, but I decided against that change eventually (but didn't reverse it fully).