Update the way pcoin_keys works to use a folder/directory instead of a single file
This is preparing tcoin/pcoin for the automatic registration of program accounts.
Eventually, I want program accounts to be tradeable by there being a way to regenerate a new key.
So, send the current key to whomever you want the program account to be used by, then the new owner regenerates the key
to complete the transfer. The old owner will no longer be able to use the old key to operate the new account. The key is
long and the key generation random enough for key collisions to be virtually impossible.