Added 'pcoin name' or 'pcoin -n' to get the name on the account
This is expected to help a pcoin escrow program confirm that a key supplied to it is that of the right pcoin account by checking what the output of 'pcoin name' is. In the future, it may be expected that 'view key', 'spend key' and 'transfer key' would be three different keys that would split power over the pcoin account. And each of the previously mentioned 'keys' could be m-of-n keys. Maybe some kind of social recovery feature could also be added (where other users confirm who you are so you can get your account back, but you nominate them as your 'social recovery circle' beforehand.